backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0

# ML & Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.15.0
joblib==1.3.2
scipy==1.11.4

# Utilities
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0